# services/quant/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quant Companion Local API",
    "description": "仅供本机桌面伴随应用调用的量化诊断与研究服务。",
    "version": "0.6.0"
  },
  "paths": {
...
    "/v1/stocks/{code}/diagnosis": {
      "get": {
        "summary": "Stock Diagnosis",
        "operationId": "stock_diagnosis_v1_stocks__code__diagnosis_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "X-Quant-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Quant-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/research/p3/demo": {
      "get": {
        "summary": "P3 Research Demo",
        "operationId": "p3_research_demo_v1_research_p3_demo_get",
        "parameters": [
          {
            "name": "X-Quant-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Quant-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/P3ResearchReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/research/p2/status": {
      "get": {
        "summary": "P2 Data Status",
        "operationId": "p2_data_status_v1_research_p2_status_get",
        "parameters": [
          {
            "name": "X-Quant-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Quant-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/P2DataReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/research/p3/real": {
      "get": {
        "summary": "P3 Research Real",
        "operationId": "p3_research_real_v1_research_p3_real_get",
        "parameters": [
          {
            "name": "X-Quant-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Quant-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/P3ResearchReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/research/p3/current": {
      "get": {
        "summary": "P3 Research Current",
        "operationId": "p3_research_current_v1_research_p3_current_get",
        "parameters": [
          {
            "name": "X-Quant-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Quant-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentSignalReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stocks/{code}/chart": {
      "get": {
        "summary": "Stock Chart",
        "operationId": "stock_chart_v1_stocks__code__chart_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 180,
              "minimum": 60,
              "default": 120,
              "title": "Limit"
            }
          },
          {
            "name": "X-Quant-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Quant-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockChartView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stocks/{code}/research": {
      "get": {
        "summary": "Stock Research",
        "operationId": "stock_research_v1_stocks__code__research_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "X-Quant-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Quant-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockResearchView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CurrentSignalReport": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "dataVersion": {
            "type": "string",
            "title": "Dataversion"
          },
          "modelVersion": {
            "type": "string",
            "title": "Modelversion"
          },
          "signalDate": {
            "type": "string",
            "format": "date",
            "title": "Signaldate"
          },
          "trainingStartDate": {
            "type": "string",
            "format": "date",
            "title": "Trainingstartdate"
          },
          "trainingEndDate": {
            "type": "string",
            "format": "date",
            "title": "Trainingenddate"
          },
          "trainingSampleCount": {
            "type": "integer",
            "title": "Trainingsamplecount"
          },
          "universeCount": {
            "type": "integer",
            "title": "Universecount"
          },
          "rankings": {
            "items": {
              "$ref": "#/components/schemas/CurrentSignalHolding"
            },
            "type": "array",
            "title": "Rankings"
          },
          "eligibleForDefault": {
            "type": "boolean",
            "title": "Eligiblefordefault"
          },
          "admissionReasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Admissionreasons"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Generatedat"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "status",
          "dataVersion",
          "modelVersion",
          "signalDate",
          "trainingStartDate",
          "trainingEndDate",
          "trainingSampleCount",
          "universeCount",
          "rankings",
          "eligibleForDefault",
          "admissionReasons",
          "generatedAt",
          "disclaimer"
        ],
        "title": "CurrentSignalReport"
      },
      "DiagnosisResult": {
        "properties": {
          "stock": {
            "$ref": "#/components/schemas/StockContext"
          },
          "compositeScore": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Compositescore"
          },
          "riskLevel": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "riskLabel": {
            "type": "string",
            "title": "Risklabel"
          },
          "horizonTradingDays": {
            "type": "integer",
            "title": "Horizontradingdays"
          },
          "excessReturnRankPercentile": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Excessreturnrankpercentile"
          },
          "upsideProbability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Upsideprobability"
          },
          "expectedReturnPercent": {
            "type": "number",
            "title": "Expectedreturnpercent"
          },
          "downsideRiskPercent": {
            "type": "number",
            "title": "Downsideriskpercent"
          },
          "dimensions": {
            "items": {
              "$ref": "#/components/schemas/ScoreDimension"
            },
            "type": "array",
            "title": "Dimensions"
          },
          "explanations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Explanations"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "modelVersion": {
            "type": "string",
            "title": "Modelversion"
          },
          "dataVersion": {
            "type": "string",
            "title": "Dataversion"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Generatedat"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "stock",
          "compositeScore",
          "riskLevel",
          "riskLabel",
          "horizonTradingDays",
          "excessReturnRankPercentile",
          "upsideProbability",
          "expectedReturnPercent",
          "downsideRiskPercent",
          "dimensions",
          "explanations",
          "warnings",
          "modelVersion",
          "dataVersion",
          "generatedAt",
          "simulated",
          "disclaimer"
        ],
        "title": "DiagnosisResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "P2DataReport": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "providerId": {
            "type": "string",
            "title": "Providerid"
          },
          "usageScope": {
            "type": "string",
            "title": "Usagescope"
          },
          "dataVersion": {
            "type": "string",
            "title": "Dataversion"
          },
          "startDate": {
            "type": "string",
            "title": "Startdate"
          },
          "endDate": {
            "type": "string",
            "title": "Enddate"
          },
          "indexCode": {
            "type": "string",
            "title": "Indexcode"
          },
          "universeCount": {
            "type": "integer",
            "title": "Universecount"
          },
          "tradingDays": {
            "type": "integer",
            "title": "Tradingdays"
          },
          "stockBarCount": {
            "type": "integer",
            "title": "Stockbarcount"
          },
          "financialRecordCount": {
            "type": "integer",
            "title": "Financialrecordcount"
          },
          "factorRows": {
            "type": "integer",
            "title": "Factorrows"
          },
          "factorDates": {
            "type": "integer",
            "title": "Factordates"
          },
          "featureNames": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Featurenames"
          },
          "pointInTimeEnforced": {
            "type": "boolean",
            "title": "Pointintimeenforced"
          },
          "universeSurvivorshipSafe": {
            "type": "boolean",
            "title": "Universesurvivorshipsafe"
          },
          "quality": {
            "$ref": "#/components/schemas/P2DataQuality"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Generatedat"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "status",
          "providerId",
          "usageScope",
          "dataVersion",
          "startDate",
          "endDate",
          "indexCode",
          "universeCount",
          "tradingDays",
          "stockBarCount",
          "financialRecordCount",
          "factorRows",
          "factorDates",
          "featureNames",
          "pointInTimeEnforced",
          "universeSurvivorshipSafe",
          "quality",
          "generatedAt",
          "disclaimer"
        ],
        "title": "P2DataReport"
      },
      "P3ResearchReport": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated"
          },
          "dataVersion": {
            "type": "string",
            "title": "Dataversion"
          },
          "featureNames": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Featurenames"
          },
          "walkForwardFolds": {
            "type": "integer",
            "title": "Walkforwardfolds"
          },
          "embargoTradingDays": {
            "type": "integer",
            "title": "Embargotradingdays"
          },
          "modelResults": {
            "items": {
              "$ref": "#/components/schemas/RankingMetricResult"
            },
            "type": "array",
            "title": "Modelresults"
          },
          "latestTop20": {
            "items": {
              "$ref": "#/components/schemas/PortfolioHoldingResult"
            },
            "type": "array",
            "title": "Latesttop20"
          },
          "backtest": {
            "$ref": "#/components/schemas/BacktestSummary"
          },
          "coveredConstraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Coveredconstraints"
          },
          "defaultModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultmodel"
          },
          "latestSignalDate": {
            "type": "string",
            "format": "date",
            "title": "Latestsignaldate"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Generatedat"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "status",
          "simulated",
          "dataVersion",
          "featureNames",
          "walkForwardFolds",
          "embargoTradingDays",
          "modelResults",
          "latestTop20",
          "backtest",
          "coveredConstraints",
          "defaultModel",
          "latestSignalDate",
          "generatedAt",
          "disclaimer"
        ],
        "title": "P3ResearchReport"
      },
      "PortfolioHoldingResult": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          }
        },
        "type": "object",
        "required": [
          "code",
          "score",
          "weight"
        ],
        "title": "PortfolioHoldingResult"
      },
...
      "RiskLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "RiskLevel"
      },
      "ScoreDimension": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "score",
          "summary"
        ],
        "title": "ScoreDimension"
      },
...
      "StockChartView": {
        "properties": {
          "stock": {
            "$ref": "#/components/schemas/StockContext"
          },
          "dataVersion": {
            "type": "string",
            "title": "Dataversion"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "title": "Startdate"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "title": "Enddate"
          },
          "trend": {
            "$ref": "#/components/schemas/TechnicalDirection"
          },
          "trendLabel": {
            "type": "string",
            "title": "Trendlabel"
          },
          "trendSummary": {
            "type": "string",
            "title": "Trendsummary"
          },
          "supportPrice": {
            "type": "number",
            "title": "Supportprice"
          },
          "resistancePrice": {
            "type": "number",
            "title": "Resistanceprice"
          },
          "latestRsi14": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestrsi14"
          },
          "latestMacdHistogram": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestmacdhistogram"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/KlinePoint"
            },
            "type": "array",
            "title": "Points"
          },
          "patterns": {
            "items": {
              "$ref": "#/components/schemas/TechnicalPattern"
            },
            "type": "array",
            "title": "Patterns"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "stock",
          "dataVersion",
          "startDate",
          "endDate",
          "trend",
          "trendLabel",
          "trendSummary",
          "supportPrice",
          "resistancePrice",
          "latestRsi14",
          "latestMacdHistogram",
          "points",
          "patterns",
          "disclaimer"
        ],
        "title": "StockChartView"
      },
      "StockContext": {
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "StockContext"
      },
...
      "TechnicalDirection": {
        "type": "string",
        "enum": [
          "bullish",
          "bearish",
          "sideways"
        ],
        "title": "TechnicalDirection"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}